# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dept": {
            "get": {
                "description": "部门分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptIndex",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/service.PageResult"
                        }
                    }
                }
            },
            "post": {
                "description": "新增部门",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptStore",
                "parameters": [
                    {
                        "description": "部门信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Dept"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dept/bind_director": {
            "post": {
                "description": "新增部门",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptBindDirector",
                "parameters": [
                    {
                        "description": "绑定参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.BindDirectorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dept/bind_manager": {
            "post": {
                "description": "新增部门",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptBindManager",
                "parameters": [
                    {
                        "description": "绑定参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.BindManagerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/dept/{id}": {
            "get": {
                "description": "单个部门",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptShow",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "新增部门",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "部门信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Dept"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "新增部门",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptDestroy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/dept/{id}/tree": {
            "get": {
                "description": "新增部门",
                "tags": [
                    "Dept 部门"
                ],
                "summary": "部门",
                "operationId": "DeptDisplayTree",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "部门ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Dept": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/types.DateTime"
                },
                "dept_name": {
                    "type": "string"
                },
                "director_id": {
                    "description": "部门主管",
                    "type": "integer"
                },
                "html": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "manager_id": {
                    "description": "部门经理",
                    "type": "integer"
                },
                "pid": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                },
                "updated_at": {
                    "$ref": "#/definitions/types.DateTime"
                }
            }
        },
...
        "req.BindDirectorReq": {
            "type": "object",
            "properties": {
                "dept_id": {
                    "type": "integer"
                },
                "director_id": {
                    "type": "integer"
                }
            }
        },
        "req.BindManagerReq": {
            "type": "object",
            "properties": {
                "dept_id": {
                    "type": "integer"
                },
                "manager_id": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "service.Links": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "prev": {
                    "type": "string"
                }
            }
        },
        "service.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "service.PageResult": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "List of data"
                },
                "links": {
                    "$ref": "#/definitions/service.Links"
                },
                "meta": {
                    "$ref": "#/definitions/service.Meta"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "types.DateTime": {
            "type": "object",
            "properties": {
                "time.Time": {
                    "type": "string"
                }
            }
        }
    }
}